pyproject.toml
[tool.poetry]
name = "TeenStudy"
version = "0.2.3"
description = "基于nonebot2异步框架的青年大学自动提交插件基于nonebot2的青年大学习自动提交插件，用于自动完成大学习，在后台留下记录，返回完成截图"
authors = ["TeenStudyFlow <TeenStudyFlow@gmail.com>"]
license="MIT"
readme="README.md"
homepage="https://github.com/YouthLearning/TeenStudy"
packages = [
  { include = "TeenStudy" },
]
[tool.poetry.dependencies]
python = "^3.8"
amis-python="^1.0.8"
anti-useragent="^1.0.10"
beautifulsoup4="^4.11.2"
bs4="^0.0.1"
fastapi="^0.100.0,<1.0.0"
httpx="^0.24.0,<1.0.0"
Jinja2="^3.1.2"
lxml="^4.9.2"
nonebot-adapter-onebot="^2.2.3"
nonebot-plugin-apscheduler="^0.3.0"
nonebot2="^2.0.0"
ddddocr="^1.4.7"
python-jose="^3.3.0"
tortoise-orm="^0.19.3"
ujson="^5.7.0"
uvicorn="^0.21.0,<1.0.0"
qrcode="^7.4.2"
psutil="^5.9.4"
pycryptodome="^3.17"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
